+++ Project/3DThinker/3dthinker/stage1/envs/trl/trl/trainer/ddpo_trainer.py
@@ -143,7 +143,7 @@ class DDPOTrainer(PyTorchModelHubMixin):
         if self.accelerator.mixed_precision == "fp16":
             inference_dtype = torch.float16
         elif self.accelerator.mixed_precision == "bf16":
-            inference_dtype = torch.bfloat16
+            inference_dtype = torch.float16
         else:
             inference_dtype = torch.float32
