+++ Project/3DThinker/3dthinker/stage1/envs/trl/trl/trainer/rloo_trainer.py
@@ -1132,7 +1132,7 @@ class RLOOTrainer(Trainer):
             ):
                 # Cast to the appropriate dtype based on training configuration
                 if self.args.bf16:
-                    unwrapped_model.to(torch.bfloat16)
+                    unwrapped_model.to(torch.float16)
                 elif self.args.fp16:
                     unwrapped_model.to(torch.float16)
                 with torch.inference_mode():
