+++ Project/3DThinker/3dthinker/stage1/src/test.py
@@ -33,7 +33,7 @@ cache_dir = args.cache_dir
 os.environ['HF_HOME'] = cache_dir
 
 processor = AutoProcessor.from_pretrained(args.load_model_path, trust_remote_code=True, cache_dir=cache_dir)
-model = Qwen2_5_VLForConditionalGeneration.from_pretrained(args.load_model_path, device_map="auto", torch_dtype=torch.bfloat16)
+model = Qwen2_5_VLForConditionalGeneration.from_pretrained(args.load_model_path, device_map="auto", torch_dtype=torch.float16)
 
 processor.tokenizer.add_tokens("<|latent_pad|>", special_tokens=True)
 processor.tokenizer.add_tokens("<|latent_start|>", special_tokens=True)
