+++ Project/3DThinker/3dthinker/stage1/transformers/src/transformers/commands/chat.py
@@ -390,7 +390,7 @@ class ChatArguments:
         metadata={
             "help": "Override the default `torch.dtype` and load the model under this dtype. If `'auto'` is passed, "
             "the dtype will be automatically derived from the model's weights.",
-            "choices": ["auto", "bfloat16", "float16", "float32"],
+            "choices": ["auto", "float16", "float16", "float32"],
         },
     )
     trust_remote_code: bool = field(
