+++ Project/3DThinker/3dthinker/stage1/transformers/src/transformers/models/bamba/convert_mamba_ssm_checkpoint.py
@@ -222,7 +222,7 @@ def convert_mamba_ssm_checkpoint_file_to_huggingface_model_file(
     state_dict = convert_state_dict_from_mamba_ssm(state_dict)
 
     # Save new model to pytorch_dump_path
-    dtype = torch.float32 if precision == "fp32" else (torch.bfloat16 if precision == "bf16" else torch.float16)
+    dtype = torch.float32 if precision == "fp32" else (torch.float16 if precision == "bf16" else torch.float16)
 
     save_file_fn = None
     if isinstance(save_model, bool) and save_model:
