+++ Project/3DThinker/3dthinker/stage1/transformers/src/transformers/models/fuyu/convert_fuyu_model_weights_to_hf.py
@@ -92,7 +92,7 @@ def convert_fuyu_checkpoint(pytorch_dump_folder_path, ada_lib_path, pt_model_pat
     state_dict = rename_state_dict(state_dict)
 
     transformers_config = FuyuConfig()
-    model = FuyuForCausalLM(transformers_config).to(torch.bfloat16)
+    model = FuyuForCausalLM(transformers_config).to(torch.float16)
     model.load_state_dict(state_dict)
     model.save_pretrained(pytorch_dump_folder_path, safe_serialization=safe_serialization)
     transformers_config.save_pretrained(pytorch_dump_folder_path)
